+++ kamehouse-shell/bin/common/csv/csv-kamehouse-shell.sh
@@ -3,9 +3,7 @@
 # Returns kamehouse-shell scripts as csv as a relative path from kamehouse-shell
 
 source ${HOME}/programs/kamehouse-shell/bin/common/functions/common-functions.sh
-if [ "$?" != "0" ]; then
-  echo "`date +%Y-%m-%d' '%H:%M:%S` - [ERROR] - Error importing common-functions.sh" ; exit 99
-fi
+if [ "$?" != "0" ]; then echo "`date +%Y-%m-%d' '%H:%M:%S` - [ERROR] - Error importing common-functions.sh" ; exit 99 ; fi
 
 initKameHouseShellEnv() {
   LOG=DISABLED