+++ kamehouse-shell/bin/common/functions/kamehouse/kamehouse-functions.sh
@@ -1,7 +1,5 @@
 source ${HOME}/programs/kamehouse-shell/bin/common/functions/common-functions.sh
-if [ "$?" != "0" ]; then
-  echo "`date +%Y-%m-%d' '%H:%M:%S` - [ERROR] - Error importing common-functions.sh" ; exit 99
-fi
+if [ "$?" != "0" ]; then echo "`date +%Y-%m-%d' '%H:%M:%S` - [ERROR] - Error importing common-functions.sh" ; exit 99 ; fi
 
 export PATH="${HOME}/programs/apache-maven/bin:${PATH}"