+++ kamehouse-shell/bin/kamehouse/deploy-kamehouse.sh
@@ -1,24 +1,16 @@
 #!/bin/bash
 
 source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/kamehouse-functions.sh
-if [ "$?" != "0" ]; then
-  echo "`date +%Y-%m-%d' '%H:%M:%S` - [ERROR] - Error importing kamehouse-functions.sh" ; exit 99
-fi
+if [ "$?" != "0" ]; then echo "`date +%Y-%m-%d' '%H:%M:%S` - [ERROR] - Error importing kamehouse-functions.sh" ; exit 99 ; fi
 
 source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/docker-functions.sh
-if [ "$?" != "0" ]; then
-  echo "`date +%Y-%m-%d' '%H:%M:%S` - [ERROR] - Error importing docker-functions.sh" ; exit 99
-fi
+if [ "$?" != "0" ]; then echo "`date +%Y-%m-%d' '%H:%M:%S` - [ERROR] - Error importing docker-functions.sh" ; exit 99 ; fi
 
 source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/build-functions.sh
-if [ "$?" != "0" ]; then
-  echo "`date +%Y-%m-%d' '%H:%M:%S` - [ERROR] - Error importing build-functions.sh" ; exit 99
-fi
+if [ "$?" != "0" ]; then echo "`date +%Y-%m-%d' '%H:%M:%S` - [ERROR] - Error importing build-functions.sh" ; exit 99 ; fi
 
 source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/deployment-functions.sh
-if [ "$?" != "0" ]; then
-  echo "`date +%Y-%m-%d' '%H:%M:%S` - [ERROR] - Error importing deployment-functions.sh" ; exit 99
-fi
+if [ "$?" != "0" ]; then echo "`date +%Y-%m-%d' '%H:%M:%S` - [ERROR] - Error importing deployment-functions.sh" ; exit 99 ; fi
 
 initKameHouseShellEnv() {
   LOAD_KAMEHOUSE_SECRETS=true