+++ kamehouse-shell/bin/lin/vlc/vlc-stop.sh
@@ -3,14 +3,10 @@
 # Fix websockets reconnecting constantly
 
 source ${HOME}/programs/kamehouse-shell/bin/common/functions/common-functions.sh
-if [ "$?" != "0" ]; then
-	echo "`date +%Y-%m-%d' '%H:%M:%S` - [ERROR] - Error importing common-functions.sh" ; exit 99
-fi
+if [ "$?" != "0" ]; then echo "`date +%Y-%m-%d' '%H:%M:%S` - [ERROR] - Error importing common-functions.sh" ; exit 99 ; fi
 
 source ${HOME}/programs/kamehouse-shell/bin/common/functions/vlc/vlc-functions.sh
-if [ "$?" != "0" ]; then
-  echo "`date +%Y-%m-%d' '%H:%M:%S` - [ERROR] - Error importing vlc-functions.sh" ; exit 99
-fi
+if [ "$?" != "0" ]; then echo "`date +%Y-%m-%d' '%H:%M:%S` - [ERROR] - Error importing vlc-functions.sh" ; exit 99 ; fi
 
 initScriptEnv() {
   DEFAULT_VLC_PORT="8080"