+++ kamehouse-shell/bin/win/shutdown/suspend.sh
@@ -1,9 +1,7 @@
 #!/bin/bash
 
 source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/kamehouse-functions.sh
-if [ "$?" != "0" ]; then
-  echo "`date +%Y-%m-%d' '%H:%M:%S` - [ERROR] - Error importing kamehouse-functions.sh" ; exit 99
-fi
+if [ "$?" != "0" ]; then echo "`date +%Y-%m-%d' '%H:%M:%S` - [ERROR] - Error importing kamehouse-functions.sh" ; exit 99 ; fi
 
 mainProcess() {
   cmd.exe "/c rundll32.exe powrprof.dll,SetSuspendState 0,1,0"